+++ eNMS/models/services/configuration/napalm_rollback.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, ForeignKey, Integer
+from sqlalchemy import ForeignKey, Integer
 
 from eNMS.database import db
 from eNMS.forms import NapalmForm