+++ eNMS/models/services/networking/napalm_traceroute.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, ForeignKey, Integer
+from sqlalchemy import ForeignKey, Integer
 
 from eNMS.database import db
 from eNMS.fields import HiddenField, IntegerField, StringField
@@ -43,7 +43,7 @@ class NapalmTracerouteForm(NapalmForm):
     destination_ip = StringField(substitution=True)
     source_ip = StringField(substitution=True)
     ttl = IntegerField(default=255)
-    vrf = StringField()
+    vrf = StringField(label="VRF", substitution=True)
     groups = {
         "Traceroute Parameters": {
             "commands": ["destination_ip", "source_ip", "timeout", "ttl", "vrf"],