+++ eNMS/static/js/table.js
@@ -1499,7 +1499,7 @@ tables.file = class FileTable extends Table {
       </a>`,
       `
       <button
-        class="btn btn-primary"
+        class="btn btn-primary parent-filtering"
         onclick="eNMS.base.showInstancePanel('folder')"
         data-tooltip="Create New Folder"
         type="button"
@@ -1507,7 +1507,7 @@ tables.file = class FileTable extends Table {
         <span class="glyphicon glyphicon-folder-open"></span>
       </button>`,
       ` <button
-        class="btn btn-primary"
+        class="btn btn-primary parent-filtering"
         onclick="eNMS.administration.showFileUploadPanel()"
         data-tooltip="Upload Files"
         type="button"
@@ -1545,7 +1545,7 @@ tables.file = class FileTable extends Table {
         <button
           type="button"
           class="btn btn-sm btn-info"
-          onclick="location.href='/download/${row.type}/${row.path}'"
+          onclick="location.href='/download/${row.type}/${row.path.slice(1)}'"
         >
           <span class="glyphicon glyphicon-export"></span>
         </button>
@@ -1603,8 +1603,11 @@ tables.file = class FileTable extends Table {
 
   get filteringConstraints() {
     const parentFiltering = ($("#parent-filtering").val() || "true") == "true";
-    if (parentFiltering)
+    if (parentFiltering) {
       return { folder_path: folderPath, folder_path_filter: "equality" };
+    } else {
+      return {};
+    }
   }
 
   postProcessing(...args) {
@@ -1616,7 +1619,7 @@ tables.file = class FileTable extends Table {
       .selectpicker()
       .on("change", function () {
         self.table.page(0).ajax.reload(null, false);
-        $("#current-folder-path").toggle();
+        $("#current-folder-path,.parent-filtering").toggle();
       });
   }
 };